--- Server/Contract/IUnityConfiguration.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using Unity;
-
-namespace Contract
-{
-    public interface IUnityConfiguration
-    {
-        void RegisterType(IUnityContainer container);
-    }
-}
