--- Server/Facade/StudentFacade.cs
@@ -1,19 +1,0 @@
-﻿using Contract;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ChildMEF
-{
-    public class StudentFacade : IStudentFacade
-    {
-        public void GetStudent(string data)
-        {
-            //Console.WriteLine(data);
-        }
-    }
-
-    
-}
