+++ Server/Snappet.Model/Common/IComponentConfiguration.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Unity;
+
+namespace Snappet.Model
+{
+    public interface IComponentConfiguration
+    {
+        void RegisterType(IUnityContainer container);
+    }
+}
