+++ Server/Snappet.Model/DataProvider/IStudentJSONDataProvider.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Snappet.Model.DataProvider
+{
+   public interface IStudentJSONDataProvider
+    {
+        void getStudentData();
+    }
+}
