+++ Server/Snappet.Model/Properties/AssemblyInfo.cs
