+++ Server/Snappet.Model/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.Configuration.ConfigurationManager" version="4.4.1" targetFramework="net461" />
+  <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net461" />
+  <package id="Unity" version="5.8.13" targetFramework="net461" />
+</packages>
