+++ Server/Snappet.WebAPI/Properties/AssemblyInfo.cs
