+++ Server/Snappet.WebAPI/packages.config
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.AspNet.Cors" version="5.2.7" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.6" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Cors" version="5.2.7" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.6" targetFramework="net461" />
+  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.8" targetFramework="net461" />
+  <package id="Microsoft.Net.Compilers" version="2.4.0" targetFramework="net461" developmentDependency="true" />
+  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net461" />
+  <package id="System.Composition.Runtime" version="1.2.0" targetFramework="net461" />
+  <package id="System.Configuration.ConfigurationManager" version="4.4.1" targetFramework="net461" />
+  <package id="System.IO" version="4.3.0" targetFramework="net461" />
+  <package id="System.Reflection" version="4.3.0" targetFramework="net461" />
+  <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net461" />
+  <package id="System.Runtime" version="4.3.0" targetFramework="net461" />
+  <package id="Unity" version="5.8.13" targetFramework="net461" />
+  <package id="Unity.Abstractions" version="3.3.1" targetFramework="net461" />
+  <package id="Unity.AspNet.WebApi" version="5.0.15" targetFramework="net461" />
+  <package id="Unity.Container" version="5.8.13" targetFramework="net461" />
+  <package id="WebActivatorEx" version="2.2.0" targetFramework="net461" />
+</packages>
