--- Server/WebAPI/App_Start/WebApiConfig.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web.Http;
-using System.Web.Http.Cors;
-using WebAPI.Unity;
-
-namespace WebAPI
-{
-    public static class WebApiConfig
-    {
-        public static void Register(HttpConfiguration config)
-        {
-            var cors = new EnableCorsAttribute("*", "*", "*");
-            config.EnableCors(cors);
-            // Web API configuration and services
-
-            // Web API routes
-            config.MapHttpAttributeRoutes();
-
-            config.Routes.MapHttpRoute(
-                name: "DefaultApi",
-                routeTemplate: "api/{controller}/{id}",
-                defaults: new { id = RouteParameter.Optional }
-            );
-            //UnityConfigurations unity = new UnityConfigurations();
-            //unity.UnityResolver();
-        }
-    }
-}
